+++ tuskar/api/controllers/root.py
@@ -20,7 +20,6 @@ import pecan
 
 from tuskar.api.controllers import v1
 
-
 class RootController(object):
 
     v1 = v1.Controller()
