+++ tuskar/api/controllers/v1/__init__.py
@@ -11,3 +11,11 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+from tuskar.api.controllers.v1.controller import Controller
+from tuskar.api.controllers.v1.data_center import DataCenterController
+from tuskar.api.controllers.v1.flavor import FlavorsController
+from tuskar.api.controllers.v1.rack import RacksController
+from tuskar.api.controllers.v1.resource_class import ResourceClassesController
+
+__all__ = (Controller, DataCenterController, FlavorsController,
+           RacksController, ResourceClassesController)
