+++ tuskar/api/controllers/v1/types/__init__.py
@@ -1,0 +1,26 @@
+# vim: tabstop=4 shiftwidth=4 softtabstop=4
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+from tuskar.api.controllers.v1.types.base import Base
+from tuskar.api.controllers.v1.types.capacity import Capacity
+from tuskar.api.controllers.v1.types.chassis import Chassis
+from tuskar.api.controllers.v1.types.error import Error
+from tuskar.api.controllers.v1.types.flavor import Flavor
+from tuskar.api.controllers.v1.types.link import Link
+from tuskar.api.controllers.v1.types.node import Node
+from tuskar.api.controllers.v1.types.rack import Rack
+from tuskar.api.controllers.v1.types.relation import Relation
+from tuskar.api.controllers.v1.types.resource_class import ResourceClass
+
+__all__ = (Base, Capacity, Chassis, Error, Link, Node, Rack,
+           Relation, ResourceClass)
