+++ tuskar/api/controllers/v1/types/capacity.py
@@ -11,3 +11,14 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+
+import wsme
+from wsme import types as wtypes
+from tuskar.api.controllers.v1.types.base import Base
+
+class Capacity(Base):
+    """A capacity representation."""
+
+    name = wtypes.text
+    value = wtypes.text
+    unit = wtypes.text
