+++ tuskar/api/controllers/v1/types/chassis.py
@@ -11,3 +11,14 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
+
+import wsme
+from wsme import types as wtypes
+from tuskar.api.controllers.v1.types.base import Base
+from tuskar.api.controllers.v1.types.link import Link
+
+class Chassis(Base):
+    """A chassis representation."""
+
+    id = wtypes.text
+    links = [Link]
